requirements.txt
absl-py==1.4.0
alembic==1.11.1
antlr4-python3-runtime==4.8
appdirs==1.4.4
astunparse==1.6.3
attrs==22.1.0
audioread==2.1.9
-e git+https://github.com/facebookresearch/WavAugment.git@54afcdb00ccc852c2f030f239f8532c9562b550e#egg=augment
autopage==0.5.1
bitarray==2.5.1
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
category-encoders==2.5.1
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.1.1
click==8.1.3
cliff==4.3.0
cmaes==0.9.1
cmd2==2.4.3
colorama==0.4.5
colorlog==6.7.0
confection==0.0.3
convertdate==2.4.0
crepe==0.0.13
cycler==0.11.0
cymem==2.0.7
Cython==0.29.30
data==0.4
de-core-news-sm==3.4.0
decorator==5.1.1
en-core-web-sm==3.4.1
et-xmlfile==1.1.0
exceptiongroup==1.0.4
fairseq==0.12.2
flatbuffers==23.1.21
fonttools==4.33.3
funcsigs==1.0.2
future==0.18.2
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.3
h5py==3.8.0
hijri-converter==2.3.1
hmmlearn==0.2.8
holidays==0.11.3.1
hydra-core==1.0.7
idna==3.4
imageio==2.19.3
importlib-metadata==4.13.0
iniconfig==1.1.1
Jinja2==3.1.2
joblib==1.1.0
julius==0.2.6
keras==2.11.0
kiwisolver==1.4.2
korean-lunar-calendar==0.3.1
langcodes==3.3.0
latex==0.7.0
lazy-loader==0.1rc2
libclang==15.0.6.1
librosa==0.9.1
lightgbm==3.3.2
littleutils==0.2.2
llvmlite==0.36.0
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
mir-eval==0.7
mlxtend==0.22.0
murmurhash==1.0.9
mvgavg==0.1.0
networkx==2.8.4
numba==0.53.0
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.0.6
openpyxl==3.1.2
opt-einsum==3.3.0
optuna==3.0.4
outdated==0.2.1
packaging==21.3
pandas==1.4.2
pandas-flavor==0.3.0
pathy==0.7.1
patsy==0.5.2
pbr==5.11.1
Pillow==9.1.1
pingouin==0.5.2
plotly==5.11.0
pluggy==1.0.0
pooch==1.6.0
portalocker==2.4.0
preshed==3.0.8
prettytable==3.7.0
primePy==1.3
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyMeeus==0.5.12
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pytest==7.2.0
python-dateutil==2.8.2
python-speech-features==0.6
pytorchltr==0.2.1
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.6.2
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.9
sacrebleu==2.1.0
scikit-image==0.19.3
scikit-learn==1.0.2
scikit-posthocs==0.7.0
scipy==1.8.0
seaborn==0.12.2
shutilwhich==1.1.0
six==1.16.0
smart-open==5.2.1
SoundFile==0.10.3.post1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==2.0.15
srsly==2.4.5
statsmodels==0.13.2
stevedore==5.1.0
tabulate==0.8.10
tempdir==0.7.1
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
thinc==8.1.5
threadpoolctl==3.1.0
tifffile==2022.5.4
tomli==2.0.1
torch==1.12.1
torch-pitch-shift==1.2.2
torchaudio==0.12.1+cu102
torchaudio-augmentations==0.2.4
torchtext==0.14.0
torchvision==0.13.1+cu102
tqdm==4.64.1
typer==0.7.0
typing-extensions==4.4.0
urllib3==1.26.13
verstack==3.7.1
wasabi==0.10.1
wavaugment==0.2
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.14.1
xarray==2022.3.0
xgboost==1.7.5
zipp==3.14.0
